"Removed CostomCoustomDitelsCard, CustomerSearchDelegate, AddCustomerDialog, CustomerDetailView, and EditCustomerDialog classes from CustomersView file."
